Move ManualFeatureConfig out of the manual feature package
