+++ WordPress/src/test/java/org/wordpress/android/util/config/manual/ManualFeatureConfigViewModelTest.kt
@@ -11,6 +11,7 @@ import org.mockito.Mock
 import org.wordpress.android.BaseUnitTest
 import org.wordpress.android.R
 import org.wordpress.android.TEST_DISPATCHER
+import org.wordpress.android.util.config.ManualFeatureConfig
 import org.wordpress.android.util.config.RemoteConfig
 import org.wordpress.android.util.config.manual.ManualFeatureConfigViewModel.FeatureUiItem.Button
 import org.wordpress.android.util.config.manual.ManualFeatureConfigViewModel.FeatureUiItem.Feature
